requirements.txt
streamlit
rembg
opencv-python-headless
pillow
numpy